pyproject.toml
[tool.poetry]
name = "ike"
version = "0.1.0"
description = "Iterative Knowledge Extractor (IKE) - An advanced system for extracting and refining knowledge from documents using LLMs"
authors = ["Thomas Kavanagh <thomaskavanagh.dev@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
Django = "5.1.1"
djangorestframework = "3.15.2"
django-cors-headers = "4.4.0"
neomodel = "5.3.3"
django_neomodel = "0.0.2"
gunicorn = "23.0.0"
streamlit = "1.38.0"
plotly = "5.24.1"
watchdog = "^4.0.0"
python-dotenv = "^1.0.1"
griptape = "0.32.0"

[tool.poetry.group.dev.dependencies]
pytest = "8.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
start = "python manage.py runserver"
migrate = "python manage.py migrate"
dashboard = "streamlit run dashboard/app.py"

[tool.pytest.ini_options]
pythonpath = ["."]
addopts = "-v -ra -q"
testpaths = ["tests"]

[tool.black]
line-length = 88
target-version = ['py39']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 88